Refactoring some state usage Containing state with component
