+++ constants/variables.js
@@ -64,3 +64,5 @@ export const TOKENS = {
 };
 
 export const REPAY_MARGIN = 1.000011;
+
+export const MESSAGE = `Hello from the Union team. Please verify your email and wallet ownership by signing this message. This doesn't cost anything and your email won't be publicly visible.`;
