+++ states/Application/index.js
@@ -1,6 +1,4 @@
 export {
-  useEmailModalOpen,
-  useEmailModalToggle,
   useGetInvitedModalOpen,
   useGetInvitedModalToggle,
   useLearnMoreModalOpen,
