+++ states/Application/states.js
@@ -4,8 +4,6 @@ export const walletModalViewState = newRidgeState("CREATE");
 
 export const walletModalState = newRidgeState(false);
 
-export const emailModalState = newRidgeState(false);
-
 export const getInvitedModalState = newRidgeState(false);
 
 export const learnMoreModalState = newRidgeState(false);
